// src/app/model/activity.ts
export interface Activity {
  id?: string;
  clientId?: string;
  title?: string;
  status?: string;
  accountStatus?: string;
  claim?: string;
  pendingTransaction?: string;
  businessProspect?: string;
  nextMeetingDate?: Date;
  agentId?: string;
  agentName?: string;
  createdDate?: Date;
  lastUpdatedDate?: Date;
  lastUpdatedBy?: string;
  meetingReport?: string;
}